readproc not needed, fixed memgraph
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky@92 7f574dfc-610e-0410-a909-a81674777703

+++ conky.c
@@ -2822,7 +2822,7 @@ static void draw_string(const char *s)
 	memcpy(tmpstring1, s, TEXT_BUFFER_SIZE);
 }
 
-inline unsigned long do_gradient(unsigned long first_colour, unsigned long last_colour) { /* this function returns the next colour between two colours for a gradient */
+inline unsigned long long do_gradient(unsigned long long first_colour, unsigned long long last_colour) { /* this function returns the next colour between two colours for a gradient */
 	int tmp_color = 0;
 	int red1, green1, blue1; // first colour
 	int red2, green2, blue2; // second colour
@@ -2876,7 +2876,7 @@ inline unsigned long do_gradient(unsigned long first_colour, unsigned long last_
 	return tmp_color;
 }
 
-inline unsigned long gradient_max(unsigned long first_colour, unsigned long last_colour) { /* this function returns the max diff for a gradient */
+inline unsigned long long gradient_max(unsigned long long first_colour, unsigned long long last_colour) { /* this function returns the max diff for a gradient */
 	int red1, green1, blue1; // first colour
 	int red2, green2, blue2; // second colour
 	int red3 = 0, green3 = 0, blue3 = 0; // difference
